refactor(codex): codex/mod.rs 拆六个语义兄弟模块
- 3120→654 行；doctor_family/thread_lifecycle/control_ops/shared_control/commit_message_gen/title_metadata
- 逐字搬运，75 个 pub(crate) 消费面经 mod.rs 重导出穷尽
- 仅必要可见性提升（4 fn + SharedControlResponseRoute 字段，供 codex_tests 消费）
- 17 个未归类命令留 mod.rs
- baseline 重生成 + 新文件登记豁免
验证：cargo check 零 error；cargo test --lib codex 341 passed；rustfmt 过；gate 绿
（子代理 CodexModSplit 执行，主代理复核验证）
