+++ docs/architecture/large-file-baseline.json
@@ -1,5 +1,5 @@
 {
-  "generatedAt": "2026-08-30T15:51:56.466Z",
+  "generatedAt": "2026-08-30T16:01:44.526Z",
   "scope": "fail",
   "policyVersion": "2026-08-30.policy-v5",
   "entries": [
